[#205] Fix deadlock issue with ydb_tp_st() in case of TP restarts
A test C program with one or more threads each of which use ydb_tp_st() to do TP transactions
deadlocked once in a while when the database has concurrent udpates that caused the TP
transaction to restart.

Below is the C-stack of the hung TP worker thread.

Thread 5 (Thread 0x7f5d1d7ad700 (LWP 94507)):
 #0  __lll_lock_wait () at ../sysdeps/unix/sysv/linux/x86_64/lowlevellock.S:135
 #1  __GI___pthread_mutex_lock (mutex=0x7f5d222252c0 <thread_mutex>) at ../nptl/pthread_mutex_lock.c:78
 #2  rts_error_va (csa=0x0, argcnt=1, var=0x7f5d1d7ac5a0) at sr_unix/rts_error.c:146
 #3  rts_error_csa (csa=0x0, argcnt=1) at sr_unix/rts_error.c:101
 #4  ydb_tp_s_common (stapi=1, tptoken=15090, tpfn=0x55c68a782bc0 <gvnset>, tpfnparm=0x7f5d2000fed0, transid=0x0, namecount=0, varnames=0x0) at sr_unix/ydb_tp_s_common.c:223
 #5  ydb_tp_sst (tptoken=15090, tpfn=0x55c68a782bc0 <gvnset>, tpfnparm=0x7f5d2000fed0, transid=0x0, namecount=0, varnames=0x0) at sr_unix/ydb_tp_sst.c:32
 #6  ydb_stm_tpthreadq_process (curTPWorkQHead=0x7f5d1001ba40) at sr_unix/ydb_stm_tpthread.c:125
 #7  ydb_stm_tpthread (parm=0x0) at sr_unix/ydb_stm_tpthread.c:80
 #8  start_thread (arg=0x7f5d1d7ad700) at pthread_create.c:463
 #9  clone () at ../sysdeps/unix/sysv/linux/x86_64/clone.S:95

(gdb) f 4
 #4  ydb_tp_s_common (stapi=1, tptoken=15090, tpfn=0x55c68a782bc0 <gvnset>, tpfnparm=0x7f5d2000fed0, transid=0x0, namecount=0, varnames=0x0) at sr_unix/ydb_tp_s_common.c:223
223                     INVOKE_RESTART;

It notices a restartable situation and therefore wants to signal that but in order to do that it
goes through the function rts_error_va() (frame #2) which tries to do a PTHREAD_MUTEX_LOCK_IF_NEEDED
and that is hung because the pthread mutex lock is held by some other thread.

Turns out this thread was obtained by this same thread in a prior restart (in rts_error_va) but
it was never released. Since all SimpleAPI and SimpleThreadAPI calls to rts_error_va() go through
ydb_simpleapi_ch(), that is where we should be releasing the pthread lock. We do that already but
towards the end of the function.  And in case of a TP restart (ERR_TPRETRY error code), we return
before reaching that point.

The main fix is to sr_unix/ydb_simpleapi_ch.c to move the PTHREAD_MUTEX_UNLOCK_IF_NEEDED macro to
the beginning of ydb_simpleapi_ch(). The pthread unlock logic that was already there did not use
the macro (code duplication) and is now removed.

While at this, various other issues were noticed and fixed.

* sr_unix/ydb_stm_thread.c
  1) It did not initialize the posix_timer_thread_id to be the worker thread.
  2) It did not initialize the thread_mutex global variable (which is what is used by the
     PTHREAD_MUTEX_LOCK_IF_NEEDED and PTHREAD_MUTEX_UNLOCK_IF_NEEDED macros). This is now done
     using a new INITIALIZE_THREAD_MUTEX_IF_NEEDED macro.

* sr_unix/gtm_multi_thread.c : This is the only place where the thread_mutex global variable was
  previously initialized. This now invokes the new INITIALIZE_THREAD_MUTEX_IF_NEEDED macro.

* sr_port/gtm_malloc_src.h : was_holder was being initialized incorrectly before invoking the
  PTHREAD_MUTEX_UNLOCK_IF_NEEDED macro just before issuing a MEMORY error. The consequences of this
  issue are that the pthread mutex lock would not be released by gtm_malloc() in case of a MEMORY error
  in case the call is done by when multi-threading is turned on. Since multi-threading was turned on
  only for MUPIP JOURNAL ROLLBACK/RECOVER, this is a non-issue until now when multi-threading is used
  due to support of ydb_malloc_t() in the SimpleThreadAPI.

* sr_unix/gtm_multi_thread.h : Defines the new INITIALIZE_THREAD_MUTEX_IF_NEEDED macro.