[#387] Use POST_BUILD and TARGET to prevent rebuilding of debug symbols
Previously, make && make install caused the debug symbols to be overwritten
by a copy of the already-stripped executable, resulting in a loss of the
debug symbols. This change prevents the generation of these symbols
from being a seperate target, which stops the make install step from
remaking the targets and copying them again.