[#392] ydb_linktmpdir/gtm_linktmpdir env var should default to ydb_tmp/gtm_tmp before defaulting to /tmp
It is a copy-paste regression introduced in GT.M V6.3-002 as part of changes to gtm_env_init_sp.c.

If TRANS_LOG_NAME returns SS_NORMAL and trans.len is > 0, then it implies the env var is defined and so
we should not consider it as an env var unspecified case. Instead what was meant is to check for
trans.len == 0 which implies the env var is undefined. Now fixed.