[#205] Fix ydb_lock_incr_st() issue with passing nanosecond timeout to ydb_lock_incr_s() on 32-bit YottaDB build (ARM); Add hex codes in comments for YDB_TP_RESTART/YDB_TP_ROLLBACK/YDB_NODE_END/YDB_LOCK_TIMEOUT/YDB_NOTOK etc. in libyottadb.h
The 64-bit nanosecond timeout passed as the first parameter to ydb_lock_incr_st() is stored
as the first 2 parameters in the 32-bit callblk.args[] array. So in ydb_stm_threadq_process(),
callblk->args[0] and callblk->args[1] should be used as the least-significant and most-significant
32-bit values respectively (instead of callblk->args[1] and callblk->args[2] respectively).

Using the wrong callblk->args[] array parameters leads to the actual timeout passed to
ydb_lock_incr_s() being completely different from the original timeout passed to ydb_lock_incr_st()
causing the simplethreadapi/lockst subtest to fail on 32-bit ARM platform with the following
signature (timeout is 0 seconds instead of the expected 1 second).

44c44
< Lock timeout test PASSED for ydb_lock_incr_st() : Timeout expected = 1 second. Actual timeout = 1 seconds
---
> Lock timeout test PASSED for ydb_lock_incr_st() : Timeout expected = 1 second. Actual timeout = 0 seconds