[#205] Enhance assert to take TP worker thread into account; Note down intrpt_ok_state in local variable before asserting on it
If asserts involving global variable "intrpt_ok_state" fail, it is possible that the actual value of the
global variable at the time of the assert failure changes before a core file is dumped. To avoid this,
we note down the global into a local variable and assert on it. The core file will keep the local
variable value intact.