[#205] Block external signals from being delivered to the TP worker thread (fix bug that intended to block previously but did not)
The SIGPROCMASK macro that was previously being in turn invoked the INSIDE_THREADED_CODE macro to
determine whether sigprocmask() or pthread_sigmask() needed to be called. But that macro was true only
if the "multi_thread_in_use" global variable was TRUE which is not the case for SimpleThreadAPI. So
we were incorrectly invoking sigprocmask() even though we were in a multi-threaded environment.