[#205] Use thread-safe system malloc() instead of ydb_malloc() in ydb_stm_getcallblk(); Removes need for a thread level lock in MAIN SimpleThreadAPI worker thread
Currently malloc() (i.e. ydb_malloc()) invocations from ydb_stm_getcallblk() are the only YottaDB
engine function invocations that can concurrently happen in other threads while the MAIN or TP
worker threads are running.  And since ydb_malloc() is not multi-thread safe, there was a need to
get a thread-level lock in the MAIN and TP worker threads. This lock already exists in the MAIN
worker thread but not in the TP worker thread but it is needed since ydb_tp_s_common (invoked from
the TP worker thread) can invoke malloc().

But instead of adding a thread-level lock in the TP worker thread, this commit changes ydb_stm_getcallblk()
to use the system malloc() which is multi-thread safe. This means the YottaDB engine will always run
only in one thread (the MAIN worker thread or the TP worker thread). And so there is no need for any
thread-level locks in either of the threads. This means the lock/release sequence in the MAIN worker
thread can also be (and is now) removed.  Should improve performance of SimpleThreadAPI.