[#387] By default build without debug symbols; CMAKE_BUILD_TYPE==Release is noticeably faster than CMAKE_BUILD_TYPE==RelWithDebInfo
We ran the following test case on an existing database.

	for i=1:1:100000 s ^x(i)=$j(i_$j,200)

A build with CMAKE_BUILD_TYPE==Release as the default turned out to be almost 5% faster than a build
with CMAKE_BUILD_TYPE==RelWithDebInfo as the default. Since speed is critical, the default is now
switched back to Release like it was before any of the prior #387 commits.

But the user still has the option to do a cmake build with CMAKE_BUILD_TYPE==RelWithDebInfo to generate
the .dbg files which should help them better debug errors/cores.