Enhance $ydb_dist/ydb_env_set script to create /tmp/yottadb and subdirectories with group and world write permissions
This is needed for the GoWrapper.

Ideally the ydb_env_set script should create with group and world write permissions based on
whether execution of YottaDB is restricted to a specific group or not. A separate issue will
be created to address that. This is an interim fix until then.