[#395] Ensure /tmp/yottadb/$ydb_ver has read-write-execute permissions for all users permitted to run YottaDB
The permissions of the temporary directory /tmp/yottadb/$ydb_ver created by ydb_env_set
(e.g. /tmp/yottadb/r124_x86_64 for YottaDB r1.24 on x86_64 platform) are read-write-execute for user,
group and world if YottaDB is installed to be world-executable, and read-write-execute for the user,
and the group permitted to run YottaDB, if YottaDB is installed to allow only members of a group
to execute it.