[#398] Fix ydb_node_previous_s() issues with global and local variables
In case of local variables, if x=1 and x(1)=2 was set and ydb_node_previous_s() was done
with x(1) as input, it used to SIG-11. This is because of a missed "return" in
sr_unix/sapi_return_subscr_nodes.c which caused the code to fall through to a for loop
that tried to interpret the TREF(sapi_query_node_subs_cnt) special value of -1 as a huge
positive number resulting in a SIG-11.

In case of global variables, the special value of -1 was never set resulting in YDB_NODE_END
being returned prematurely. The -1 setting now happens in sr_unix/sapi_save_targ_key_subscr_nodes.c
just like it is done for local variables in sr_port/op_fnreversequery.c.