[#205] Fix more cases of not displaying correct SimplethreadAPI or SimpleAPI function that user invoked in case of SIMPLEAPINEST and other errors
A new macro LYDBRTNNAME is introduced that is passed the lydbrtntype (e.g. LYDB_RTN_GET)
as a parameter and it returns "ydb_get_s()" in case of SimpleAPI or "ydb_get_st()" in case of
SimpleThreadAPI.

All error invocations use this macro so the user visible error message indicates the correct
function that the user invoked (ydb_get_s or ydb_get_st etc.).

This involved reworking the table in libyottadb_rtns.h. The 2nd column there was removed as it
was never used and instead a new column was added to note down the SimpleThreadAPI routine name.

While at this, various existing usages were optimized to use the new LYDBRTNNAME macro.

For example, the following usage

	simpleThreadAPI_active ? "ydb_get_st()" : "ydb_get_s()"

is changed to

	 LYDBRTNNAME(LYDB_RTN_GET)