[#205] Issue SIMPLEAPINEST error in SimpleThreadAPI mode and avoid deadlocks; Various misc fixes
The core change is in sr_unix/ydb_stm_args.c to detect if we are the MAIN worker thread and if so
not queue a request for service (since we are about to deadlock) and instead issue a SIMPLEAPINEST
error where appropriate.

And if a new TP transaction is being started while inside a ydb_ci_t() or ydb_cip_t() we need to
know the exact caller and so the "calltyp" is now being passed to "ydb_cip_helper" so it can set
TREF(libyottadb_active_rtn) to LYDB_RTN_YDB_CI or LYDB_RTN_YDB_CIP.

And instead of shutting off the active routine indicator in "ydb_cip_helper" for the entire duration
of the call-in, it is now moved to only for the duration of any SimpleThreadAPI request that is
serviced by the MAIN worker thread in sr_unix/ydb_stm_args.c.

And related to this, the SIMPLEAPINEST error message was simplified to use !AZ (null-terminated
strings) instead of !AD which meant we needed to do a STRLEN (since the length was not readily
available) before passing the parameters in the setup_error() function call.

While at this, the LIBYOTTADB_INIT_* macros in sr_unix/libyottadb_int.h were simplified.

Also the 3rd column for LYDB_RTN_TP was replaced from "********" to "ydb_tp_st()" since this was
now being used to issue the SIMPLEAPINEST error in sr_unix/ydb_stm_args.c.