[#205] Prevent SIG-11 in ydb_init() in case a STAPIFORKEXEC error is about to be returned
In case a STAPIFORKEXEC error is about to be issued, a SETUP_GENERIC_ERROR macro is invoked
to trigger the error. But as part of that macro as TREF(ydb_error_code) is done. And this
relies on a SETUP_THREADGBL_ACCESS having been invoked but that is done only a little
after the STAPIFORKEXEC error codepath in ydb_init(). This is now moved to the beginning
of ydb_init() much before we are about to issue the STAPIFORKEXEC error.