Use gcc LTO (LinkTimeOptimization) when building YottaDB; Reduces executable size and improves run times
The core change is to sr_linux/platform.cmake where CMAKE_C_FLAGS now uses -flto to enable
Link Time Optimization feature of gcc. In addition it uses -fuse-ld=gold to use the gold linker
which is found to be slightly better than the default ld linker. These changes are currently enabled
only for Release or RelWithDebInfo builds, not for Debug builds of YottaDB.

Using -flto identified more warnings that are now fixed. One example of the kind of warning was
a global variable which was defined with a GBLDEF specifying one type whereas it was declared with a
GBLREF in various other files using a different type. These were till now not visible to the linker
but -flto makes all the different C files visible to the linker thereby identifying more warnings.
Another warning was function prototypes. All functions used in the transfer table (in sr_port/xfer.h
and sr_port/xfer_table.c) were currently declared with a prototype indicating a "int" return value
whereas this was not true for a lot of the functions. All of them are now fixed by explicitly including
the header file that has the prototype (for C functions) and declaring a prototype explicitly in
xfer_table.c (for assembly functions).

README.md is enhanced to indicate a potential issue while building in Ubuntu 18.10 (due to a "binutils"
package bug that has been since fixed) and a workaround.