[#205] Add extra parameter <errstr> to SimpleThreadAPI so error string is always accurate in a multi-threaded environment
This address the issue where a multi-threaded program invokes some SimpleThreadAPI function
(e.g. ydb_get_st) and gets an error returned back. And then invokes ydb_zstatus() to see the error
string corresponding to the returned error but since ydb_zstatus() looks at a process-wide-global
ISV $zstatus, it could correspond to an error from another thread that also got an error in the
YottaDB engine soon after our ydb_get_st() call.

The fix is to pass in a "errstr" parameter to all SimpleThreadAPI calls. This is a ydb_buffer_t
pointer which can be NULL in which case no error string is filled in. If errstr is non-NULL, and
an error is encountered during a SimpleThreadAPI invocation (i.e. ydb_simpleapi_ch is invoked),
errstr->buf_addr is filled with the $zstatus string (upto errstr->len_alloc bytes including a
terminating null byte which means it could be truncated too) by ydb_simpleapi_ch(). This avoids
the concurrency issue since at the time ydb_simpleapi_ch() is invoked, the MAIN worker thread is
running the YottaDB engine and servicing only one request even though the process is multi-threaded.

A new global variable TREF(stapi_errstr) is set to point to the user-passed errstr parameter
in the MAIN worker thread (ydb_stm_thread.c) before it invokes a SimpleAPI function. This is
relied upon by ydb_simpleapi_ch.c as an indication to fill in the error string. TREF(stapi_errstr)
is reset to NULL soon after the request is serviced in the MAIN worker thread.