[#401] Fix SIG-11 and/or non-deterministic behavior when a call-in invokes an external call that uses Simple API
The primary fix is to sr_port/alias_funcs.c. The SKIP_BASE_FRAMES macro invocation was done towards
the end of the do/while loop. If "fp" at the beginning of the do/while loop is a call-in base frame
(possible if this is a SimpleAPI invocation), then we would be doing "fp = fp->old_frame_pointer"
BEFORE we do the SKIP_BASE_FRAMES macro which meant it was possible for fp to become NULL at the
time the SKIP_BASE_FRAMES macro is invoked which would result in it not skipping like it would
have done if it was invoked with a non-null "fp". The fix is to move the SKIP_BASE_FRAMES macro to
the beginning of the do/while loop just like it is done in various other places in the code base
(sr_port/zlput_rname.c, sr_unix/gtm_trigger.c, sr_unix/relinkctl.c).

A similar fix was needed in sr_port/op_clralsvars.c. It was doing something similar but only
for the call-in base frames (whereas SKIP_BASE_FRAMES does it for call-in and trigger base frames).
To avoid code duplication, the SKIP_BASE_FRAMES macro is now enhanced to take in a second parameter
which can be (SFT_CI | SFT_TRIGR) in all current callers and will be (SFT_CI) in op_clralsvars.c.
This lets us use the same code to check for call-in/trigger as well as just call-in base frames.
This meant changing all existing callers of SKIP_BASE_FRAMES to pass in a second parameter of
(SFT_CI | SFT_TRIGR).

There was a check of "fp" against stackbase/stacktop in op_clralsvars.c and alias_funcs.c which
seemed unnecessary so have converted it to a debug-only assert (IS_PTR_INSIDE_M_STACK macro).