[#205] [#358] Regenerate GTMDefinedTypesInit*.m due to SimpleThreadAPI and LIBAIO changes
The SimpleThreadAPI changes (#205) apply to sr_x86_64/sr_armv7l/sr_aarch64.

The LIBAIO changes (#358) apply only to sr_armv7l/sr_aarch64.

While at this, fixed the CACHELINE_PAD macro to not generate a semi-colon and changed all
macro callers to instead insert the semi-colon at the end to be more consistent with other
similar macro usages. This is a cosmetic change.