[#205] Preserve signal handler information passed by OS (e.g. sending pid etc.) before forwarding signal to another thread
This ensures a SIG-3 sent by a pid shows up as KILLBYSIGUINFO message (which corresponds to an
externally generated signal and has sending pid details) instead of a KILLBYSIGSINFO1 message
(which is an internally generated signal in the process and has no sending pid details).

A new global variable "exi_signal_forwarded" records a non-zero value (the signal number) in case
forwarding happened in generic_signal_handler(). And records the incoming "info" and "context" from
the OS into global variables "exi_siginfo" and "exi_context". When the signal gets forwarded and
generic_signal_handler() gets invoked again, we use this new global variable to detect that this is
a forwarded situation and use the global variables as is instead of using the input "info" and
"context" in that signal handler invocation. This way we preserve the original signal handler
information even in case of a forwarded signal. Only signals in generic_signal_handler() currently
benefit from this scheme as all other signal handlers in YottaDB do not use either "info" or "context".
But this change meant a change to the FORWARD_SIG_TO_MAIN_THREAD_IF_NEEDED macro interface and so
all callers had to change a bit.