[AARCH64] ydbinstall.sh should support YottaDB 64-bit ARM only on Ubuntu 18.04 and higher
A prior commit added support on Ubuntu 18.04 and higher for 64-bit ARM but the change was done
at a location which meant it also supported RHEL on 64-bit ARM (because that was code lying
around for the x86_64 platform).

This is now fixed so the 64-bit ARM architecture has separate code and only Ubuntu 18.04 is
allowed there.