Fix few more build warnings from gcc LTO; Speed up LTO time by parallelizing link
The LTO (Link Time Optimization) build of gcc showed a few warnings only on an Ubuntu 16.04 system.
The warnings were that a global variable was declared with inconsistent types across multiple C files.
In all these cases, the GBLDEF type was taken as the correct one and all the GBLREF usages were
fixed to match the GBLDEF. This also meant in some cases adding a type cast before calling system
functions (e.g. rename(), unlink() etc.).

Since the global variable "source_name_len" was fixed to be an "unsigned short" everywhere, for
consistency the global variable "object_name_len" was also fixed to be an "unsigned short" everywhere.

In addition, the cmake script now determines the # of CPUs available on the system and passes it
to the -flto option to let the link also be parallelized. This greatly cut the link time for
libyottadb.so (in one system, total build time reduced from 3.25 minutes to 1.75 minutes).