[#205] In ydb_fork_n_core/gtm_fork_n_core, preserve C-stack of all threads in core for better debugging (only for DEBUG builds)
A prior commit enabled a similar change in sr_unix/ch_cond_core.c but that was done only in case of
a fatal error in the YottaDB engine. It is possible for ydb_fork_n_core/gtm_fork_n_core to be called
without the engine encountering a fatal error (e.g. a test C program that uses the SimpleThreadAPI
could encounter a YDB_ASSERT macro failure which will invoke ydb_fork_n_core) and we want the
C-stack of all threads even in that case for better debugging.

Note that because of this change, the code flow for PRO vs DBG builds is different. In PRO, one
would invoke a ydb_fork_n_core/gtm_fork_n_core, generate a core (to create a snapshot of the process
state for later debugging) but the process would continue. Whereas in DBG, the process would create
the core and terminate right then. Given this is done only in DBG builds, it is considered okay.