[#205] Handle potential YDB_ERR_CALLINAFTERXIT return from TP callback function in TP worker thread
The v60000/gtm4525b subtest failed (1 out of 100 runs or so) with the following assert failure.
The TP worker thread asked for a op_trollback to be done by the MAIN worker thread (using the
LYDB_RTN_TP_ROLLBACK_TLVL0 opcode) but instead got a YDB_ERR_CALLINAFTERXIT status returned.
This is possible because the test does a MUPIP STOP (i.e. kill -15) of processes which would
cause the MAIN/TP worker threads to be signaled to terminate at which point, any more requests
from them will return with YDB_ERR_CALLINAFTERXIT. The assert is now modified to take this into
account. Although this did not show up in the test failure, a similar issue exists with the opcode
LYDB_RTN_TP_ROLLBACK_TLVL0 and so even in that case we now handle the possibility of
YDB_ERR_CALLINAFTERXIT.

Below is a gdb session of the core failure for the record.

 #0 pthread_kill () from /usr/lib64/libpthread.so.0
 #1 gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2 ch_cond_core () at sr_unix/ch_cond_core.c:76
 #3 rts_error_va () at sr_unix/rts_error.c:194
 #4 rts_error_csa () at sr_unix/rts_error.c:101
 #5 ydb_stm_tpthreadq_process () at sr_unix/ydb_stm_tpthread.c:259
 #6 ydb_stm_tpthread (parm=0x0) at sr_unix/ydb_stm_tpthread.c:83
 #7 start_thread () from /usr/lib64/libpthread.so.0
 #8 clone () from /usr/lib64/libc.so.6

(gdb) f 5
 #5  ydb_stm_tpthreadq_process () at sr_unix/ydb_stm_tpthread.c:259
259                                                     assert(YDB_TP_ROLLBACK == rlbk_retval);

(gdb) p rlbk_retval
$1 = -150381530

(gdb) p int_retval
$2 = -150381530

libydberrors.h:#define YDB_ERR_CALLINAFTERXIT -150381530