[#205] Make ydb_init() multi-thread safe in case of STAPIFORKEXEC and CALLINAFTERXIT errors; Ensure ydb_zstatus() returns error string after a CALLINAFTERXIT return from ydb_init
Ensure we get the YottaDB engine pthread mutex before checking for STAPIFORKEXEC or CALLINAFTERXIT
errors. This is needed since SETUP_GENERIC_ERROR macro (which is invoked in both these error scenarios)
operates on global variables (e.g. dollar_zstatus) and therefore has to be multi-thread safe since
ydb_init() is supposed to be multi-thread safe.

Invoke SETUP_GENERIC_ERROR macro in case of a CALLINAFTERXIT error in ydb_init(). This error was being
issued in two places in the same function. The duplication is now removed too.