[#205] If exit handling is deferred in SimpleThreadAPI process (e.g. TP transaction in final retry when SIGTERM/SIG-15 is received), MAIN/TP worker threads should service requests without YDB_ERR_CALLINAFTERXIT errors until the TP transaction commits
If SIGTERM is sent and generic_signal_handler() gets invoked, if we find that DEFER_EXIT_PROCESSING
is TRUE, we do not invoke the exit handler right away but instead defer it until it is safe to
start exit handler processing. But we do invoke SET_FORCED_EXIT_STATE which would set the global
variable "forced_thread_exit" to TRUE. And since this is the variable currently relied upon by the
SimpleThreadAPI worker threads (ydb_stm_thread.c and ydb_stm_tpthread.c), they would return
YDB_ERR_CALLINAFTERXIT on all pending requests in their work queues. But this means that if a TP
transaction is active and in the final retry (which also means we are holding crit on the database)
at the time the SIGTERM was received and exit handling was deferred, this TP transaction will never
commit fine because any ydb_*_s() requests done inside this callback function after the SIGTERM signal
got sent would return with a YDB_ERR_CALLINAFTERXIT error. This is not desirable as we want the
crit-holding transaction to be done as soon as possible and in a clean fashion.

Therefore, the worker threads design is reworked a bit to now rely on "forced_simplethreadapi_exit",
a new global variable. On seeing this, they will exit right away (what they previously used to do
when they saw "forced_thread_exit" to be TRUE).

ydb_exit() and generic_signal_handler() will invoke SET_FORCED_EXIT_STATE to set "forced_thread_exit"
to TRUE whenever they want the SimpleThreadAPI process to terminate at the next logical point.
The MAIN worker thread, before it attempts to service any request, will check "forced_thread_exit"
and if it finds this to be TRUE, but finds "forced_simplethreadapi_exit" to be FALSE and OK_TO_INTERRUPT
is TRUE, it will set "forced_simplethreadapi_exit" to be TRUE to indicate the logical point has been
reached and that the worker thread should exit right away.