[#205] If MAIN worker thread gets SIG-15, wait for MAIN/TP worker threads to reach logical point before starting exit handler processing
We had a test failure (in the dual_fail_extend/dual_fail2_mustop_sigquit subtest) where a SimpleThreadAPI
process was sent a SIG-15 by the test and the signal got delivered to the MAIN worker thread but it
went ahead with exit handler processing (including rolling back an active TP transaction) while a
TP worker thread was concurrently running the TP callback function without realizing all of this going on.
The TP worker thread effectively got an INVTPTRANS error since it was using a non-zero tptoken in a
ydb_set_st() call when there was no active TP transaction (due to the exit handler doing an op_trollback()).

The fix is to defer exit processing in generic_signal_handler.c if we find out that we are the
MAIN worker thread. This way the MAIN worker thread will invoke the exit handler gtm_exit_handler()
inside ydb_stm_thread() when it knows it is a logical/safe point to do so.

In addition, deferred_signal_handler() is now fixed to skip invoking the exit handler in case we
are the MAIN worker thread. This is because ydb_stm_thread() has an already established scheme
(using "forced_simplethreadapi_exit" global variable) to determine the logical point and then invoke
gtm_exit_handler().

Below is the C-stack of all threads at the time of the core for the record.

(gdb) thread apply all bt

Thread 3 (Thread 0x7fde4cb67700 (LWP 14698)):
 #0  fsync () from /usr/lib64/libc.so.6
 #1  jnl_fsync (reg=0x55af6c90e7b8, fsync_addr=38517184) at sr_unix/jnl_fsync.c:134
 #2  wcs_flu (options=519) at sr_unix/wcs_flu.c:413
 #3  gds_rundown (cleanup_udi=1) at sr_unix/gds_rundown.c:608
 #4  gv_rundown () at sr_port/gv_rundown.c:123
 #5  gtm_exit_handler () at sr_unix/gtm_exit_handler.c:216
 #6  __run_exit_handlers () from /usr/lib64/libc.so.6
 #7  exit () from /usr/lib64/libc.so.6
 #8  gtm_image_exit (status=-15) at sr_unix/gtm_image_exit.c:27
 #9  generic_signal_handler (sig=15, info=0x7fde4cb66830, context=0x7fde4cb66700) at sr_unix/generic_signal_handler.c:380
 #10 <signal handler called>
 #11 pthread_cond_wait@@GLIBC_2.3.2 () from /usr/lib64/libpthread.so.0
 #12 ydb_stm_thread (parm=0x0) at sr_unix/ydb_stm_thread.c:123
 #13 start_thread () from /usr/lib64/libpthread.so.0
 #14 clone () from /usr/lib64/libc.so.6

Thread 2 (Thread 0x7fde510c6dc0 (LWP 14695)):
 #0  do_futex_wait.constprop () from /usr/lib64/libpthread.so.0
 #1  __new_sem_wait_slow.constprop.0 () from /usr/lib64/libpthread.so.0
 #2  ydb_stm_args (callblk=0x55af6c96b550) at sr_unix/ydb_stm_args.c:183
 #3  ydb_stm_args5 (tptoken=0, errstr=0x0, calltyp=16, p1=94211928230125, p2=140733677288928, p3=94211928265280, p4=1, p5=140733677288912) at sr_unix/ydb_stm_args.c:320
 #4  ydb_tp_st (tptoken=0, errstr=0x0, tpfn=0x55af6c8408ed <tpfn_stage1>, tpfnparm=0x7fff1cd7bde0, transid=0x55af6c849240 <tptypebuff> "BATCH", namecount=1, varnames=0x7fff1cd7bdd0) at sr_unix/ydb_tp_st.c:33
 #5  impjob (childnum=2) at simplethreadapi_imptp.c:1148
 #6  main (argc=1, argv=0x7fff1cd7c198) at simplethreadapi_imptp.c:602

Thread 1 (Thread 0x7fde47fff700 (LWP 14705)):
 #0  pthread_kill () from /usr/lib64/libpthread.so.0
 #1  gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  ch_cond_core () at sr_unix/ch_cond_core.c:76
 #3  rts_error_va (csa=0x0, argcnt=7, var=0x7fde47ffeaa0) at sr_unix/rts_error.c:194
 #4  rts_error_csa (csa=0x0, argcnt=7) at sr_unix/rts_error.c:101
 #5  ydb_stm_args (callblk=0x7fde40000b20) at sr_unix/ydb_stm_args.c:126
 #6  ydb_stm_args4 (tptoken=7085, errstr=0x0, calltyp=12, p1=94211928265184, p2=2, p3=94211928261632, p4=94211928263568) at sr_unix/ydb_stm_args.c:298
 #7  ydb_set_st (tptoken=7085, errstr=0x0, varname=0x55af6c8491e0 <ygbl_arandom>, subs_used=2, subsarray=0x55af6c848400 <subscr>, value=0x55af6c848b90 <ybuff_val>) at sr_unix/ydb_set_st.c:33
 #8  tpfn_stage1 (tptoken=7085, errstr=0x0, parm_array=0x7fff1cd7bde0) at simplethreadapi_imptp.c:1384
 #9  ydb_stm_tpthreadq_process (curTPWorkQHead=0x7fde48024c40, forced_simplethreadapi_exit_seen=0x7fde47ffeea8) at sr_unix/ydb_stm_tpthread.c:225
 #10 ydb_stm_tpthread (parm=0x0) at sr_unix/ydb_stm_tpthread.c:84
 #11 start_thread () from /usr/lib64/libpthread.so.0
 #12 clone () from /usr/lib64/libc.so.6